Reset stars dimensions when calculating differences
There have been upstream changes in stars that throw errors when doing
calculations with objects that have the same dimensions but different
values for those dimensions. This happens to us when calculating changes
in depths because we subset two different dates of depths and difference
them. This change resets the time dimension to allow these calculations.