Fix max counts for level=subregion
The use of .data was causing errors due to use of `$` when specifying single column
names not being allowed and likely further down the pipe as well. This removes use
of .data in the section of the code triggerd by level = "subregion".

It also fixes a problem caused by using bind_rows with different named columns
("region" in one table, "subregion" in another table).

The goal was to maintain the behavior of the function as written, but I think
there is also a conversation to be had about how level = "subregion" results in
data source mixing that may be unintuitive to users and should be clearly
documented.